updated dashboard page to fit admin infos
